refactor(node): NodeId usize→u32 代际化（20bit index + 12bit gen + INVALID）
T1 of v1.3+ 动态树重构：NodeId(pub u32) + impl NodeId（index/gen/INVALID/is_valid）。
全代码 .0 当 usize 索引处加 as usize（T1 临时过渡，T2/T3 改 slotmap API）；
NodeId(N) 字面量/N 的 usize 变量加 as u32；FFI NodeId(node_id) 对齐 u32 签名。
cargo build --all-targets 全过；node_id_ 3 新测 + 全 workspace 既有测全过。

+++ loomgui_core/examples/dump_interact.rs
@@ -14,7 +14,7 @@ fn dump(stage: &Stage, label: &str, focus: u32) {
     // 取最近 tick 的 frame：重跑 build_render_nodes 不现实；改读 scene 节点状态 + prev_hashes
     println!("=== {} ===", label);
     for (i, n) in scene.nodes.iter().enumerate() {
-        if i == focus || n.parent == Some(loomgui_core::scene::node::NodeId(focus)) {
+        if i == focus || n.parent == Some(loomgui_core::scene::node::NodeId(focus as u32)) {
             println!(
                 "  n{} kind={:?} active={} hovered={} rect=({:.0},{:.0},{:.0},{:.0}) bg={:?} wm={:?}",
                 i,
@@ -41,7 +41,7 @@ fn dump_frame(stage: &mut Stage, label: &str, focus: u32) {
         let nid = rn.node_id as usize;
         if nid == focus_us
             || (nid < scene.nodes.len()
-                && scene.nodes[nid].parent == Some(loomgui_core::scene::node::NodeId(focus_us)))
+                && scene.nodes[nid].parent == Some(loomgui_core::scene::node::NodeId(focus_us as u32)))
         {
             let pk = match &rn.payload {
                 NodePayload::Mesh { verts, colors, .. } => {
